python-utils/requirements.txt
flask
whisper_timestamped
python-dotenv
